kinematic-model: rise default bikerPower
Try to give more realistic time estimates by adapting the default bikerPower
to the given default totalMass. This should hopefully result in reasonable time
estimates for an average (untrained) biker. However, it's of course better to
adapt the kinematic model parameters on an indiviaul basis to get more precise
results.